Update transaction page data display
