+++ packages/frontend/src/app/validator/[hash]/ValidatorPage.scss
@@ -1,7 +1,6 @@
 @use '../../../styles/mixins.scss';
 @import '../../../styles/variables.scss';
 
-
 @mixin blockOffsets ($property) {
   #{$property}: 12px;
